+++ Tests/Unit/Domain/Model/Event/EventDateTest.php
@@ -1312,6 +1312,36 @@ final class EventDateTest extends UnitTestCase
     /**
      * @test
      */
+    public function isHybridForOnSiteEventReturnsFalse(): void
+    {
+        $this->subject->setEventFormat(EventDateInterface::EVENT_FORMAT_ON_SITE);
+
+        self::assertFalse($this->subject->isHybrid());
+    }
+
+    /**
+     * @test
+     */
+    public function isHybridForOnlineEventReturnsFalse(): void
+    {
+        $this->subject->setEventFormat(EventDateInterface::EVENT_FORMAT_ONLINE);
+
+        self::assertFalse($this->subject->isHybrid());
+    }
+
+    /**
+     * @test
+     */
+    public function isHybridForHybridEventReturnsTrue(): void
+    {
+        $this->subject->setEventFormat(EventDateInterface::EVENT_FORMAT_HYBRID);
+
+        self::assertTrue($this->subject->isHybrid());
+    }
+
+    /**
+     * @test
+     */
     public function getWebinarUrlInitiallyReturnsEmptyString(): void
     {
         self::assertSame('', $this->subject->getWebinarUrl());